backend/fashion-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
faiss-cpu>=1.7.4
Pillow>=9.5.0
python-multipart>=0.0.6
pymongo>=4.6.0
python-dotenv>=1.0.0